requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
crispy-bootstrap5==0.7
django==4.2
django-crispy-forms==2.0
django-user-agents==0.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
pytz==2023.3
setuptools==67.7.1 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
tzdata==2023.3
ua-parser==0.16.1
user-agents==2.2.0
whitenoise==6.4.0
